+++ mortar-sample/src/main/java/com/example/mortar/view/FriendView.java
@@ -31,8 +31,8 @@ public class FriendView extends TextView  {
     Mortar.inject(context, this);
   }
 
-  @Override protected void onAttachedToWindow() {
-    super.onAttachedToWindow();
+  @Override protected void onFinishInflate() {
+    super.onFinishInflate();
     presenter.takeView(this);
   }